getdents_calls

+++ src/trusted/service_runtime/nacl_syscall_strace.c
@@ -41,11 +41,10 @@ void NaClStraceDup3(int oldfd,int newfd,int flags,int ret){
     printf("dup3(%d, %d, %d) = %d\n", oldfd, newfd, flags, ret);
 
 }
-void NaClStraceGetdents(int d,void *drip,int  count,size_t ret,ssize_t getdents_ret,uintptr_t sysaddr){
-    printf("getdents(%d, %p, %d) = %zu, Getdents Ret: %zd, Sysaddr: %p\n", d, drip, count, ret, getdents_ret, (void*)sysaddr);
-
-
+void NaClStraceGetdents(int d, void *drip, size_t count, size_t ret, ssize_t getdents_ret, uintptr_t sysaddr) {
+    printf("getdents(%d, %p, %zu) = %zu, Getdents Ret: %zd, Sysaddr: %p\n", d, drip, count, ret, getdents_ret, (void *)sysaddr);
 }
+
 void NaClStracePread(int d, void *buf, int count,  size_t log_bytes,int32_t ret){
     printf("pread(%d, %p, %d, %zu) = %d\n", d, buf, count, log_bytes, ret);
 }
